+++ frontend/src/hooks/useProducts.ts
@@ -1,53 +1,5 @@
-import { useState, useEffect } from "react";
-import axios from "axios";
-import header from "../components/Header.tsx";
-import {config} from "typescript-eslint";
-
-interface Product {
-  name: string;
-  price: number;
-  country_of_origin: string;
-  description: string;
-  fruit_or_vegetable: string;
-  expiry_date: string;
-  imageId: string;
-  file?: string;
-}
-
-interface ProductResponse {
-  product?: Product;
-}
-
-const fetchProducts = async (): Promise<Product[]> => {
-  try {
-    const response = await axios.get<ProductResponse>(
-      `http://localhost:8002/api/products/6775934ed79a82364c118356`,
-        // {headers: {Access-Control-Allow-Origin: "*"}}
-
-
-    );
-    console.log(response.data);
-    console.log(response.data.product);
-    return   [response.data.product]
-
-    // return ;
-
-    // if (product) {
-    //   const imageResponse = await axios.get<Blob>(
-    //     `http://localhost:8000/api/products/6772091ea203351edda37ccf/image`,
-    //     { responseType: "blob" },
-    //   );
-    //   const imageUrl = URL.createObjectURL(imageResponse.data);
-    //   product.imageUrl = imageUrl;
-    //   return [product];
-    // } else {
-    //   return [];
-    // }
-  } catch (error) {
-    console.error("Error fetching products:", error);
-    return [];
-  }
-};
+import { useEffect, useState } from "react";
+import { fetchProducts } from "../api/productsApi.ts";
 
 const useProducts = () => {
   const [products, setProducts] = useState<Product[]>([]);
