+++ frontend/src/model/product.ts
@@ -1,0 +1,10 @@
+interface Product {
+    name: string;
+    price: number;
+    country_of_origin: string;
+    description: string;
+    fruit_or_vegetable: string;
+    expiry_date: string;
+    imageId: string;
+    file?: string;
+}
